refactor: initServer to separate module
